// client/web/src/schema/settings.schema.d.ts
export interface InsightDashboard {
  /**
   * Title of the dashboard.
   */
  title: string;
  id: string;
  /**
   * Insights ids that will be included in the dashboard.
   */
  insightIds?: string[];
  [k: string]: unknown;
}